Add a seekable argument to the read_file_handle_callable methods. (#22)
If true, this forces the backend to return a seekable handle.  The short version is that this means we do not use streaming, and we have to buffer the contents.  This increases our memory consumption by ~2x, but it should be survivable.

Add a test for loading numpy arrays over http.  Fails without the changes in slicedimage/{backends/*,io.py}, but passes with the changes.  Yay!