Convert read_file_handle{,_callable} to read_contextmanager (#40)
Context managers checks the feature list we need for this interface: an object that promises a file-like read handle, but doesn't do the work until it's invoked.  Previously, we did this by returning a method that, when called, would return the file-like handle.  The downside is that we can't execute any cleanup logic.  With context managers, we can require that a block of code is executed once the file-like read handle is no longer necessary.

Currently, this is just used to close the file handle (by invoking the wrapped context managers's __exit__method), but the next step is to use this cleanup block to verify that the data read matched the checksums.