Prevent cache pollution (#53)
Somehow, I made a mistake in merging my changes, and did not pass the checksum to the backend (I swear I ran into the same problem last time and fixed it already, but clearly it's not in the repo...).  To prevent repeating the mistake, I wrote a test to cover this case.

Bump the cache version to invalidate the caches with this bug.