Rename tests/io to tests/io_ (#89)
For reasons that are not clear to me, python is seemingly picking up tests/io _before_ the system io module now.  It didn't seem to do this in the past but is now.  Furthermore, it only reproduces on travis.  As a test, I picked a previous run of master that succeeded, and re-ran it, and it failed!

With this fix, it seems to work.  <shrug>