Delay S3 resource construction until object requested (#111)
Constructing a S3 handle takes up memory and time.  We retrieve the backend whenever we initialize a tile, but we want to delay the expensive work until the data is requested.

The perf issue is sufficiently bad that we want to do a release, so bumping to 4.0.1.

Test plan: Able to quickly load an experiment where all the tiles are located in the cloud.