Fix Windows path resolution (#112)
When extracting the path component from urllib.parse.urlparse, the path always starts with '/'.  When tokenizing the url into its components, it becomes '/', 'c:', 'windows', 'system32'.  Unfortunately, when combined into a WindowsPath, it becomes 'c:windows/system32', which honestly makes no sense to me.

The approach here is to strip the prefix '/', and when we do that, WindowsPath can parse the posix path correctly.

Test plan: Ran this on a Windows VM.