[easy] Reverse context unwind (#119)
We have a mechanism for tracking multiple context managers for unit test setup and teardown.  For unwinding them context managers, we should unwind them in the reverse order they are created.

On POSIX systems, this does not matter, as we can hold file handles to files that are deleted.  However, on Windows, the ordering is critical as one cannot delete files that are still held open.

In the case where we create a temporary directory, and then spin up a http server to listen to requests at that directory, we need to kill the process _before_ we clean up the temporary directory.

Test plan: works on windows
Part of https://github.com/spacetx/starfish/issues/1296